requirements.txt
sklearn
tensorflow==1.12.0
keras==2.2.0
tensorflow-estimator==1.10.12
tensorflow-serving-api==1.12.0
numpy
pillow
scikit-image
opencv-python
pandas
jupyter
sagemaker
Cython
contextlib2
lxml
matplotlib 
mysqlclient
mysql
pymysql

